chore(models): refactor models to individual files
