+++ models/brand_tier.go
@@ -1,0 +1,9 @@
+package models
+
+type BrandTier string
+
+const (
+	BrandTierPremium BrandTier = "premium"
+	BrandTierRegular BrandTier = "regular"
+	BrandTierBudget  BrandTier = "budget"
+)
