+++ models/cart_item.go
@@ -1,0 +1,7 @@
+package models
+
+type CartItem struct {
+	Product  Product `json:"product"`
+	Quantity int     `json:"quantity"`
+	Size     string  `json:"size"`
+}
