+++ models/customer_profile.go
@@ -1,0 +1,6 @@
+package models
+
+type CustomerProfile struct {
+	ID   string `json:"id"`
+	Tier string `json:"tier"`
+}
