+++ models/payment_info.go
@@ -1,0 +1,14 @@
+package models
+
+type PaymentInfo struct {
+	Method   string  `json:"method"`
+	BankName *string `json:"bank_name"`
+	CardType *string `json:"card_type"`
+}
+
+func (p *PaymentInfo) BankNameOrDefault() string {
+	if p.BankName != nil {
+		return *p.BankName
+	}
+	return ""
+}
