+++ models/product.go
@@ -1,0 +1,14 @@
+package models
+
+import (
+	"github.com/shopspring/decimal"
+)
+
+type Product struct {
+	ID           string          `json:"id"`
+	Brand        string          `json:"brand"`
+	BrandTier    BrandTier       `json:"brand_tier"`
+	Category     string          `json:"category"`
+	BasePrice    decimal.Decimal `json:"base_price"`
+	CurrentPrice decimal.Decimal `json:"current_price"`
+}
